Extract InteractiveSession from CoreReplApp.Interactive partial
Move interactive REPL session management, ambient command dispatch,
history handling, prompt building, and input tokenization into a
standalone InteractiveSession class. Execution calls Interactive
through the app facade, breaking the bidirectional partial coupling.
